package_support: add *flags() functions; prefer files to 'export -f'.
The use of "set -x" has been reduced because a number of loops spammed too
much.
The calls of sed to extract package descriptions and fill the .yypkg.script
files with package values have been replaced with a single awk script which
is not vulnerable to the use of sed's delimiter in variables and thus
doesn't require double-escaping everything.