deps: fix libocaml build.
With current ocaml versions, we need to provide a "-I ." for configure to
find some of its required files. We do so through the $(CONFIGURE_ARGS)
variable. However we also need it to appear before other arguments so we
move the occurrence of $(CONFIGURE_ARGS) earlier in the command-line used
to invoke the compiler.